Add client.rename_dont_move()
This patch adds another `mv`-primitive that is purposefully limited to
only rename and not moving of files. While this seldom is a problem when
interactively using the `mv` command. It often causes problems in
production environments when multiple programs do something like:

    mv a dest
    mv b dest  # Should throw error, but creates "dest/b" instead

Also, we're hoping to forever fix the plethora of luigi issues related
to atomicity in renaming that have arised:

 * https://github.com/spotify/luigi/pull/393
 * https://github.com/spotify/luigi/pull/522
 * https://github.com/spotify/luigi/pull/557

And also the tons of tickets on Spotify's internal JIRA.

Luigi also has a PR to integrate this once this is merged.

 * https://github.com/spotify/luigi/pull/605