FIX: Fix formatting of permissions with sticky bit
Snakebite always prepended a leading 0 to permissions when using the 'stat'
command. This led to invalid permissions being returned. For example in cases
like "1777" it returned "01777". This is supposed to fix this case.

To ensure compatibility with python3 octal numbers used in
stat_test are switched to representation with leading '0o'
defined in pep-3127

In python3 oct() returns 0o prefix, so the method for formatting
needs to be adjusted for compatibility too.