Minor fix for doc of client.AutoConfigClient
The current documentation states that HADOOP_PATH is used to determine additional
locations of configuration files. Instead HADOOP_HOME (and actually HADOOP_CONF_DIR, but
with this one I'm not sure if its consistently supported) is used. Also, it's worth
mentioning that core-site.xml is required for configuration as well to determine
default file-system name.