Make kerberos/sasl support optional
Since installing the python-krbV dependency requires the kerberos
libraries to be installed, this commit makes kerberos/sasl support
optional at time of installation.

To install snakebite with sasl/kerberos support run:

  pip install "snakebite[kerberos]"

This triggers the installation of the sasl and python-krbV
dependendies.

In the client, sasl and kerberos are only included when the
client is explicitly configured with SASL.