Increase client robusteness by adding retry and request timeout
Add a request retry logic for connection and request failures. Mimics
the behavior of the Java Hadoop client's FailoverOnNetworkExceptionRetry
strategy. Uses the retry and failover values from the Hadoop
configuration if set. This enables the client to not fail in case the
NameNode RPC queues are full or other transient failures happened that
cannot be solved by one failover.

Configure the channel socket to use the socket timeout as configured
for Hadoop after establishing the connection. This helps to not
time out when running slow operations on the NameNode, such as
large listings, or when having large and filled up RPC queues.