Support dfs.client.use.datanode.hostname config property
When reading blocks, client talks to datanode by default using the datanode's
IP address returned by the name node. The problem with this approach is that
the IP address is a private IP within the cluster, so the client won't works if
it is in a different network. This commit updates the config module to support
dfs.client.use.datanode.hostname config property, and have the clients to use
this property to determine weather to use the IP address or the hostname to
talk to datanodes.