Improve python3 support
In python 3 long and int types have been unified and it is no longer
legal to use the L postfix to declare a long.  As this postfix is used
in snakebite it fails under python 3.  This commit improves python 3
support for some use cases.

It is not intended to make snakebite fully python 3 compatible in any
way but just to remove a common issue.  This changes allows limited
functionality under python 3 which is enough for many usecases.